bot/requirements.txt
more-itertools==9.0.0
Pillow==9.2.0
pg8000==1.29.3
pytesseract==0.3.10
python-telegram-bot==20.0a4
pytz==2022.5
reportlab==3.6.11
SQLAlchemy==1.4.42
types-cachetools==5.2.0